requirements.txt
# OCR Web Application Dependencies
# Author: M Hamza Ummer
# Version: 2.0.0

# Core Flask dependencies
Flask==2.3.3
Werkzeug==2.3.7

# OCR dependencies
pytesseract==0.3.10
Pillow==10.0.1
pdf2image==1.16.3

# PDF creation (for testing)
reportlab==4.4.3

# File type detection
python-magic==0.4.27
python-magic-bin==0.4.14  # For Windows compatibility

# Development and testing
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0

# Production server (optional)
gunicorn==21.2.0

# Security
cryptography==41.0.7

# HTTP requests (for testing)
requests==2.31.0
